add classes of errors and create connroller login

+++ controllers/cards.js
@@ -1,15 +1,21 @@
 const Card = require('../models/card');
-const { ERROR_DEFAULT, ERROR_CODE, ERROR_NOT_FOUND } = require('../utils/constants/errorStatus');
+const BadRequestError = require('../errors/bad-request-error');
+const NotFoundError = require('../errors/not-found-error');
+const DefaultError = require('../errors/default-error');
+const ForbiddenError = require('../errors/forbidden-error');
 
 // GET Получить все карточки
-module.exports.getCards = (req, res) => {
+module.exports.getCards = (req, res, next) => {
   Card.find({})
     .then((cards) => res.status(200).send({ data: cards })) // статус 200, отправляем карточки
-    .catch(() => res.status(ERROR_DEFAULT).send({ message: 'На сервере произошла ошибка' })); // ошибка сервера, статус 500
+    .catch(() => {
+      throw new DefaultError('На сервере произошла ошибка');
+    })
+    .catch(next);
 };
 
 // POST Создать карточку
-module.exports.createCard = (req, res) => {
+module.exports.createCard = (req, res, next) => {
   const { name, link } = req.body;
   const owner = req.user._id;
 
@@ -17,34 +23,46 @@ module.exports.createCard = (req, res) => {
     .then((card) => res.status(201).send(card))
     .catch((err) => {
       if (err.name === 'ValidationError') {
-        res.status(ERROR_CODE).send({ message: 'Переданы некорректные данные при создании карточки.' });
-        return;
+        throw new BadRequestError('Переданы некорректные данные при создании карточки.');
       }
-      res.status(ERROR_DEFAULT).send({ message: 'На сервере произошла ошибка' });
-    });
+      throw new DefaultError('На сервере произошла ошибка');
+    })
+    .catch(next);
 };
 
 // DELETE Удалить карточку
-module.exports.deleteCard = (req, res) => {
-  Card.findByIdAndRemove(req.params.id)
+module.exports.deleteCard = (req, res, next) => {
+  Card.findById(req.params.id)
     .then((card) => {
       if (!card) {
-        res.status(ERROR_NOT_FOUND).send({ message: 'Передан несуществующий _id карточки.' });
-        return;
+        throw new NotFoundError('Передан несуществующий _id карточки.');
       }
-      res.status(200).send({ message: 'Карточка удалена' });
+      if (String(card.owner) === req.user._id) {
+        Card.findByIdAndRemove(req.params.id)
+          .then(() => {
+            res.status(200).send({ message: 'Карточка удалена' });
+          })
+          .catch((err) => {
+            if (err.name === 'CastError') {
+              throw new BadRequestError('Передан некорректный _id карточки.');
+            }
+            throw new DefaultError('На сервере произошла ошибка');
+          });
+      } else {
+        throw new ForbiddenError('Вы не являетесь создателем данной карточки');
+      }
     })
     .catch((err) => {
       if (err.name === 'CastError') {
-        res.status(ERROR_CODE).send({ message: 'Переданы некорректные данные при удалении карточки.' });
-        return;
+        throw new BadRequestError('Передан некорректный id карточки.');
       }
-      res.status(ERROR_DEFAULT).send({ message: 'На сервере произошла ошибка' });
-    });
+      throw new DefaultError('На сервере произошла ошибка');
+    })
+    .catch(next);
 };
 
 // PUT Поставить лайк карточке
-module.exports.likeCard = (req, res) => {
+module.exports.likeCard = (req, res, next) => {
   Card.findByIdAndUpdate(
     req.params.id,
     { $addToSet: { likes: req.user._id } }, // добавить _id в массив, если его там нет
@@ -52,22 +70,21 @@ module.exports.likeCard = (req, res) => {
   )
     .then((card) => {
       if (!card) {
-        res.status(ERROR_NOT_FOUND).send({ message: 'Передан несуществующий _id карточки.' });
-        return;
+        throw new NotFoundError('Передан несуществующий _id карточки.');
       }
       res.status(200).send(card);
     })
     .catch((err) => {
       if (err.name === 'CastError') {
-        res.status(ERROR_CODE).send({ message: 'Переданы некорректные данные для постановки лайка.' });
-        return;
+        throw new BadRequestError('Переданы некорректные данные для постановки лайка.');
       }
-      res.status(ERROR_DEFAULT).send({ message: 'На сервере произошла ошибка' });
-    });
+      throw new DefaultError('На сервере произошла ошибка');
+    })
+    .catch(next);
 };
 
 // DELETE Удалить лайк у карточки
-module.exports.dislikeCard = (req, res) => {
+module.exports.dislikeCard = (req, res, next) => {
   Card.findByIdAndUpdate(
     req.params.id,
     { $pull: { likes: req.user._id } }, // убрать _id из массива
@@ -75,16 +92,15 @@ module.exports.dislikeCard = (req, res) => {
   )
     .then((card) => {
       if (!card) {
-        res.status(ERROR_NOT_FOUND).send({ message: 'Передан несуществующий _id карточки.' });
-        return;
+        throw new NotFoundError('Передан несуществующий _id карточки.');
       }
       res.status(200).send(card);
     })
     .catch((err) => {
       if (err.name === 'CastError') {
-        res.status(ERROR_CODE).send({ message: 'Переданы некорректные данные для снятия лайка.' });
-        return;
+        throw new BadRequestError('Переданы некорректные данные для снятия лайка.');
       }
-      res.status(ERROR_DEFAULT).send({ message: 'На сервере произошла ошибка' });
-    });
+      throw new DefaultError('На сервере произошла ошибка');
+    })
+    .catch(next);
 };
